refactor#61: enum타입 위치 변경
