--- app/src/main/java/com/scrap/scrap2024/data/Preferences.kt
@@ -1,13 +1,0 @@
-package com.scrap.scrap2024.data
-
-enum class OrderType {
-    ASC, DESC
-}
-
-enum class SortType {
-    TITLE, SCRAP_DATE
-}
-
-enum class ViewType {
-    LIST, GRID
-}
