+++ app/src/main/java/com/scrap/scrap2024/data/enums/OrderType.kt
@@ -1,0 +1,5 @@
+package com.scrap.scrap2024.data.enums
+
+enum class OrderType {
+    ASC, DESC
+}
