+++ app/src/main/java/com/scrap/scrap2024/data/enums/SortType.kt
@@ -1,0 +1,5 @@
+package com.scrap.scrap2024.data.enums
+
+enum class SortType {
+    TITLE, SCRAP_DATE
+}
