+++ app/src/main/java/com/scrap/scrap2024/data/enums/ViewType.kt
@@ -1,0 +1,5 @@
+package com.scrap.scrap2024.data.enums
+
+enum class ViewType {
+    LIST, GRID
+}
